refactor on the page block and rework headings into its own component
